Treat strings consistently.
If a tagged template literal contains embedded expressions, we treat strings and
arrays-of-strings in expressions consistently: both are now concatenated into
the source before rendering, and only Nodes in expressions are treated specially
with placeholders. (And, we’re careful to use the same placeholder for multiple
elements in an array expression for greater efficiency.)

This means that you can embed HTML-in-HTML (or Markdown-in-Markdown, etc.)
without needing to repeat the tagged template literal, and it’s also more
efficient since we only need to render the source once.